Add files via upload
Customer churn is a financial term that refers to the loss of a client or customer—that is, when a customer ceases to interact with a company or business. Similarly, the churn rate is the rate at which customers or clients are leaving a company within a specific period of time. A churn rate higher than a certain threshold can have both tangible and intangible effects on a company’s business success. Ideally, companies like to retain as many customers as they can.

With the advent of advanced data science and machine learning techniques, it’s now possible for companies to identify potential customers who may cease doing business with them in the near future. In this article, you’ll see how a bank can predict customer churn based on different customer attributes such as age, gender, geography, and more.

This is a binary classification problem. I used python libraries and Jupyter Notebook. I got an accuracy score of 84.16% on the validation set using Logistic Regression model and Recursive Feature Elimination technique. I have uploaded all the relevant files(e.g. The Dataset: Bank Customer Churn Modeling). Thank you.